ref: extract functionality to hooks
